Refactor tests to use new builders
